requirements.txt
transformers==4.27.1
nltk~=3.8.1
torch
regex~=2022.10.31
tqdm~=4.65.0
numpy~=1.24.2
PyYAML~=6.0
openai~=0.27.2
bert-extractive-summarizer
spacy
torchvision